% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze.R
\name{Analyze}
\alias{Analyze}
\title{Analyze}
\usage{
Analyze(data, mod, stoich = "", abund = "", ref.state = NA)
}
\arguments{
\item{data}{data frame with 12 columns as decribed in documentation. Output of Normalize function}

\item{mod}{If any chemoform of a given peptide bears the modification specified in mod="" (f.ex. mod="phosphorylation") all peptides with the same sequence and containing the modification will be classified as Modified, while all peptides with the same sequence and not containing the modification will be classified as Not-Modified. If none of the chemoforms of a peptide contains the specified modification, the peptide is classified Abundance.}

\item{stoich}{user's preference for calculating the site occupancy of peptides for which either calculations are possible. Possible arguments: "Exact" or "Approximate".}

\item{ref.state}{option to normalize to a reference condition (e.g. ref.state="Disease").}

\item{adund}{user's preference for exact or approximate abundance calculations. Possible arguments: "Exact" or "Approximate".}
}
\value{
data frame of analyzed data
}
\description{
This function calculates peptides site occupancy and/or abundance.
}
\examples{
normalized <- Normalize(testData,tot.current = T,mod="Phosphorylation")
Analyze(data=normalized,mod="Phosphorylation",stoich="Exact")
}

