% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{Normalize}
\alias{Normalize}
\title{Normalization}
\usage{
Normalize(data, iso.norm = NA, equim.iso.norm = NA, internal.norm = NA,
  tot.current = F, mod)
}
\arguments{
\item{data}{data frame with 12 columns as decribed in documentation.}

\item{iso.norm}{takes a character string corresponding to the reference isotopologue to which the other isotopologue will be normalized (e.g. iso.norm = "H"). The string must match the corresponding annotation in the "Label" column.}

\item{equim.iso.norm}{takes a character string corresponding to the reference isotopologue to which the other isotopologue will be normalized (e.g. iso.norm = "H"). The string must match the corresponding annotation in the "Label" column.}

\item{internal.norm}{takes in a vector of character strings corresponding to the peptide sequence of the reference peptides to which other peptides are normalized (e.g. internal.norm = "ATDVIVP"). If more than one reference peptide is indicated (e.g. internal.norm = c("ATDVIVP","AAATDVI")), a geometric mean of the corresponding intensity values is taken. This normalization equalizes the intensities for the specified peptides in each sample.}

\item{tot.current}{takes in a boolean input indicating whether or not normalization to total ion current is to be made (e.g. tot.current = T). This normalization equalizes the sum of all intensities in each sample, used as a proxy for total ion current.}

\item{mod}{If any chemoform of a given peptide bears the modification specified in mod="" (f.ex. mod="phosphorylation") all peptides with the same sequence and containing the modification will be classified as Modified, while all peptides with the same sequence and not containing the modification will be classified as Not-Modified. If none of the chemoforms of a peptide contains the specified modification, the peptide is classified Abundance.}
}
\value{
data frame of normalized data
}
\description{
This function offers several modes of normalization to correct for different sources of error, and enables calculating differential protein expression and stoichiometry of post-tranlsational modification even from sparsely annotated datasets.
}
\examples{
Normalize(testData,tot.current = T,mod="Phosphorylation")
}

