% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QC}
\alias{QC}
\title{QC}
\usage{
QC(data, mod)
}
\arguments{
\item{data}{data frame with 12 columns as decribed in documentation.}

\item{mod}{If any chemoform of a given peptide bears the modification specified in mod="" (f.ex. mod="phosphorylation") all peptides with the same sequence and containing the modification will be classified as Modified, while all peptides with the same sequence and not containing the modification will be classified as Not-Modified. If none of the chemoforms of a peptide contains the specified modification, the peptide is classified Abundance.}
}
\value{
data frame of filtered data
}
\description{
In this quality control function, the data is checked for the presence of all specified columns in the correct order. Data is then filtered to remove peptides without any valid (i.e. < 0) intensity measurement and classified based on the possibility to perform exact or approximate stoichiometry and abundance calculations according to the modification specified by the user (e.g. "phosphorylation" or "acetylation"). The filtered and classified data is exported as a CSV file into the working directory (file name), along with a bar graph summarizing the number of peptides and proteins remaining after each step of quality control.
}
\examples{
QC(testData,mod="Phosphorylation")
}

