% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testData}
\alias{testData}
\title{Test Data for examples}
\format{An object of class \code{data.frame} with 1000 rows and 12 columns.}
\usage{
testData
}
\description{
Test Data for examples
}
\author{
Alex Kentsis \email{kentsisresearchgroup@gmail.com}
}
\keyword{testData}

